\name{ImportExportApp}
\alias{ImportExportApp}

\title{ Runs the shiny app

}
\description{ Runs a shiny app which uses all the content of the package to import and export databases in 
a rather easy way.

}
\usage{
ImportExportApp(...)
}

\arguments{
  \item{...}{ See \code{\link[shiny]{runApp}}

}
}
\details{ It requires a few packages to run the app: \pkg{shiny}, \pkg{shinyBS}, \pkg{shinythemes}, \pkg{compareGroups}.}




\seealso{
\code{\link[shiny]{runApp}}
}
\examples{
\dontrun{
ImportExportApp() }
}
