\name{ImportExport-package}
\alias{ImportExport-package}
\alias{ImportExport}
\docType{package}
\title{  \packageTitle{ImportExport} }
\description{
\packageDescription{ImportExport}

The package also includes an usefull shiny app called by \code{\link{ImportExportApp}} which uses all the content of the package to import and export databases in 
a rather easy way.

}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{ImportExport}
\packageIndices{ImportExport}
}
\author{  
\packageAuthor{ImportExport}

Maintainer: \packageMaintainer{ImportExport} 
}

\keyword{ ImportExport }
\seealso{

\code{\link{ImportExportApp}}
}
\examples{
\dontrun{
x<-spss_import("mydata.sav")
}
}
